refactor(vue-next): move pure type definition files to types directory
