+++ packages/hippy-vue-next/src/built-in-component.ts
@@ -22,7 +22,8 @@
  * Tags supported by Hippy built-in
  */
 
-import { NATIVE_COMPONENT_MAP, HIPPY_DEBUG_ADDRESS, type NeedToTyped } from './config';
+import type { NeedToTyped } from '../types';
+import { NATIVE_COMPONENT_MAP, HIPPY_DEBUG_ADDRESS } from './config';
 import { registerElement, type ElementComponent } from './runtime/component';
 import type { EventsUnionType } from './runtime/event/hippy-event';
 import { Native } from './runtime/native';
