+++ packages/hippy-vue-next/src/runtime/event/hippy-event-dispatcher.ts
@@ -21,8 +21,7 @@
 /**
  * Hippy event dispatcher, which distributes events triggered by users or generated by Native
  */
-
-import type { NeedToTyped } from '../../config';
+import type { NeedToTyped } from '../../../types';
 import { trace } from '../../util';
 import { getNodeById } from '../../util/node-cache';
 import type { HippyNode } from '../node/hippy-node';
