+++ packages/hippy-vue-next/src/runtime/event/hippy-event-target.ts
@@ -20,8 +20,7 @@
 
 import { looseEqual } from '@vue/shared';
 
-import type { CallbackType } from '../../config';
-
+import type { CallbackType } from '../../../types';
 import type { HippyEvent } from './hippy-event';
 
 export interface EventListenerOptions {
