+++ packages/hippy-vue-next/src/runtime/event/hippy-event.ts
@@ -19,7 +19,7 @@
  */
 
 // eslint-disable-next-line max-classes-per-file
-import type { NeedToTyped } from '../../config';
+import type { NeedToTyped } from '../../../types';
 import type { HippyEventTarget } from './hippy-event-target';
 
 /**
