+++ packages/hippy-vue-next/src/runtime/node/hippy-node.ts
@@ -17,15 +17,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-import type { CallbackType } from '../../config';
+import type { CallbackType, NativeNode } from '../../../types';
 import { getUniqueId, DEFAULT_ROOT_ID } from '../../util';
 import { getHippyCachedInstance } from '../../util/instance';
 import { preCacheNode } from '../../util/node-cache';
 import type { TagComponent } from '../component';
 import type { HippyEvent } from '../event/hippy-event';
 import { HippyEventTarget } from '../event/hippy-event-target';
-import type { NativeNode } from '../native/native-node';
 import {
   renderInsertChildNativeNode,
   renderRemoveChildNativeNode,
