+++ packages/hippy-vue-next/src/util/rem.ts
@@ -17,7 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import type { NeedToTyped } from '../config';
+import type { NeedToTyped } from '../../types';
 import { Native } from '../runtime/native';
 import { getHippyCachedInstance } from './instance';
