+++ packages/hippy-vue-next/types/native-modules/device-event-module.ts
