+++ packages/hippy-vue-next/types/native-modules/http.ts
@@ -17,8 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-import type { NeedToTyped } from '../../../config';
+import type { NeedToTyped } from '../index';
 
 export interface HttpRequestParams {
   method: string;
