+++ packages/hippy-vue-next/types/native-modules/image-loader-module.ts
