+++ tsconfig.json
@@ -79,5 +79,6 @@
     "packages/global.d.ts",
     "node_modules/@types/web/index.d.ts",
     "node_modules/@types/node/index.d.ts",
-  ],
+    "packages/*/types"
+  ]
 }
